+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.9]
+
+- Adds the following commands:
+  - block navigation up-left, up-right, up, down, down-left, down-right
+  - block selection
+  - navigate to the next / previous blank line
+  - scroll the editor until the cursor is in the middle of the screen
+  - configurable editor scrolling (because the built-in editor scrolling feature only goes 1 line at a time)
+- Changes nomenclature away from "interesting lines" and "boring lines" to "blocks" and "stop-lines"
+
 ## [0.2]
 
 - Default keybindings and explanation in README